bad state order in BinaryAlphabet

+++ src/Bpp/Seq/ProbabilisticSymbolList.cpp
@@ -71,10 +71,10 @@ BasicProbabilisticSymbolList::BasicProbabilisticSymbolList(const CruxSymbolList 
   alphabet_(list.getAlphabet()), content_(list.getAlphabet()->getResolvedChars().size(),list.size())
 {
   int nbc=(int)getAlphabet()->getResolvedChars().size();
-  
+
   for (size_t i=0; i<size(); i++)
     for (int s=0; s<nbc; s++)
-      content_(s,i)=list.getStateValueAt(i,(int)s);
+      content_(s,i)=list.getStateValueAt(i,s);
 }
 
 BasicProbabilisticSymbolList & BasicProbabilisticSymbolList::operator=(const ProbabilisticSymbolList & list)
